Time: 32 ms (61.26%), Space: 15.3 MB (95.37%) - LeetHub

+++ 2938-separate-black-and-white-balls/2938-separate-black-and-white-balls.cpp
@@ -1,11 +1,11 @@
 class Solution {
 public:
-    int minimumSteps(string s) {
-        int white = 0;
-        int res = 0;
-        int n = s.length();
+    long long minimumSteps(string s) {
+        long long white = 0;
+        long long res = 0;
+        long long n = s.length();
 
-        for(int i = 0 ; i < n ; ++i) {
+        for(long long i = 0 ; i < n ; ++i) {
             if(s[i] == '0') {
                 res += white;
             } else if (s[i] == '1') {
